+++ src/components/map/popups/MapPopup.vue
@@ -2,7 +2,7 @@
     <div
         class="map-popup"
         :style="{
-            '--max-height': `${overlay.maxHeight}px`,
+            '--max-height': `${overlay._maxHeight}px`,
             '--position-x': typeof overlay.position === 'object' ? `${overlay.position.x}%` : undefined,
             '--position-y': typeof overlay.position === 'object' ? `${overlay.position.y}%` : undefined,
         }"