+++ src/composables/atc.ts
@@ -1,4 +1,8 @@
-import type { VatsimShortenedController } from '~/types/data/vatsim';
+import type { VatsimExtendedPilot, VatsimShortenedAircraft, VatsimShortenedController } from '~/types/data/vatsim';
+import { getHoursAndMinutes } from '~/utils';
+import type { Map } from 'ol';
+import type { ShallowRef } from 'vue';
+import { isPilotOnGround } from '~/composables/pilots';
 
 export const useFacilitiesIds = () => {
     const dataStore = useDataStore();
@@ -37,7 +41,7 @@ export function getControllerPositionColor(controller: VatsimShortenedController
     return radarColors.neutral800;
 }
 
-export function sortControllersByPosition<T extends {facility: number, [key: string]: any}>(facilities: T[]): T[] {
+export function sortControllersByPosition<T extends { facility: number, [key: string]: any }>(facilities: T[]): T[] {
     const ids = useFacilitiesIds();
 
     const getPositionIndex = (position: number) => {
@@ -50,5 +54,43 @@ export function sortControllersByPosition<T extends {facility: number, [key: str
 
     return facilities.slice().sort((a, b) => {
         return getPositionIndex(a.facility) > getPositionIndex(b.facility) ? 1 : -1;
+    });
+}
+
+export function findAtcByCallsign(callsign: string) {
+    const dataStore = useDataStore();
+    const local = dataStore.vatsim.data.locals.value.find(x => x.atc.callsign === callsign)?.atc;
+    if (local) return local;
+
+    return dataStore.vatsim.data.firs.value.find(x => x.controller?.callsign === callsign)?.controller;
+}
+
+export function getATCTime(controller: VatsimShortenedController) {
+    return getHoursAndMinutes(new Date(controller.logon_time).getTime());
+}
+
+export function findAtcAirport(atc: VatsimShortenedController) {
+    const dataStore = useDataStore();
+    const title = atc.callsign.split('_')[0];
+    const iataAirport = dataStore.vatspy.value?.data.airports.find(x => x.iata === title);
+    if (iataAirport) {
+        return {
+            ...iataAirport,
+            icao: iataAirport.iata!,
+        };
+    }
+
+    return dataStore.vatspy.value?.data.airports.find(x => x.icao === title);
+}
+
+export function showAtcOnMap(atc: VatsimShortenedController, map: Map | null) {
+    map = map || inject<ShallowRef<Map | null>>('map')!.value;
+    const view = map?.getView();
+    const airport = findAtcAirport(atc);
+    if(!airport) return;
+
+    view?.animate({
+        center: [airport.lon, airport.lat],
+        zoom: 14,
     });
 }